Use consistent system prompt across endpoints
