+++ functions/facts.ts
@@ -2,12 +2,11 @@ import Anthropic from '@anthropic-ai/sdk';
 import { getStore } from '@netlify/blobs';
 import {
   AnthropicModel,
-  asSseStream,
+  asSseStream, currentDateSentence,
   errorResponse,
   simulateTokenGeneration,
-  storeResponse,
+  storeResponse, SYSTEM_PROMPT,
 } from '../src/lib/functions';
-import { currentDateSentence } from '../src/lib/utils';
 
 async function getWikidataId(search: string): Promise<string | undefined> {
   const baseUrl = 'https://www.wikidata.org/w/api.php';
@@ -96,10 +95,6 @@ function wikidataInfoAsCsv(wikidataInfo: Array<WikidataDatum>): string {
 
 async function formatWithClaude(search: string, wikidataInfo: Array<WikidataDatum>) {
   const client = new Anthropic({ apiKey: process.env.ANTHROPIC_API_KEY });
-  const system = `\
-You are a data retrieval assistant for the Atlas of Space, an educational tool to help learn about our solar system.
-
-Keep your responses brief. You do not have a personality, you simply state facts. ${currentDateSentence()}`;
   const prompt = `\
 What facts can you tell me about '${search}'? I'm interested in things like the following:
 
@@ -148,7 +143,7 @@ ${wikidataInfoAsCsv(wikidataInfo)}
 
   const stream = client.messages.stream({
     model: AnthropicModel.CLAUDE_4_SONNET,
-    system,
+    system: SYSTEM_PROMPT,
     messages: [{ role: 'user', content: prompt }],
     max_tokens: 1024,
   });
