+++ functions/summary.ts
@@ -3,13 +3,12 @@ import { getStore } from '@netlify/blobs';
 import {
   AnthropicModel,
   asSse,
-  asSseStream,
+  asSseStream, currentDateSentence,
   errorResponse,
   fromSseStream,
   simulateTokenGeneration,
-  storeResponse,
+  storeResponse, SYSTEM_PROMPT,
 } from '../src/lib/functions';
-import { currentDateSentence } from '../src/lib/utils';
 
 export default async function handle(request: Request) {
   const params = new URL(request.url).searchParams;
@@ -31,10 +30,6 @@ export default async function handle(request: Request) {
   }
 
   const client = new Anthropic({ apiKey: process.env.ANTHROPIC_API_KEY });
-  const system = `\
-You are a fact generation assistant for the Atlas of Space, an interactive Solar System explorer. You present facts \
-with a frank and direct tone and do not have a personality or refer to yourself in your responses. \
-${currentDateSentence()}`;
   const prompt = `\
 Generate a 1-sentence summary of ${search}. Examples of good summaries:
 
@@ -48,7 +43,7 @@ hemisphere, and by a prominent equatorial ridge that makes it resemble a walnut.
 </example>`;
   const messageStream = client.messages.stream({
     model: AnthropicModel.CLAUDE_4_SONNET,
-    system,
+    system: SYSTEM_PROMPT,
     messages: [{ role: 'user', content: prompt }],
     max_tokens: 1024,
   });
