+++ functions/visit.ts
@@ -3,13 +3,12 @@ import { getStore } from '@netlify/blobs';
 import {
   AnthropicModel,
   asSse,
-  asSseStream,
+  asSseStream, currentDateSentence,
   errorResponse,
   fromSseStream,
   simulateTokenGeneration,
-  storeResponse,
+  storeResponse, SYSTEM_PROMPT,
 } from '../src/lib/functions';
-import { currentDateSentence } from '../src/lib/utils';
 
 export default async function handle(request: Request) {
   const params = new URL(request.url).searchParams;
@@ -32,11 +31,6 @@ export default async function handle(request: Request) {
   }
 
   const client = new Anthropic({ apiKey: process.env.ANTHROPIC_API_KEY });
-  const system = `\
-You are a fact generation assistant for the Atlas of Space, an interactive Solar System explorer. You present facts \
-with a frank and direct tone and do not have a personality or refer to yourself in your responses. Keep your response \
-direct and to-the-point: do NOT preface it with information like 'Based on the available search results' or other \
-preambles. ${currentDateSentence()}`;
   const prompt = `\
 Generate a brief 1-sentence summary of ${search}. Do not restate the organization that launched the spacecraft. \
 Examples of good summaries:
@@ -57,7 +51,7 @@ and MASCOT landers before collecting a 5.4 gram sample that was returned to Eart
 </example>`;
   const messageStream = client.messages.stream({
     model: AnthropicModel.CLAUDE_4_SONNET,
-    system,
+    system: SYSTEM_PROMPT,
     messages: [{ role: 'user', content: prompt }],
     max_tokens: 1024,
     tools: [
