+++ src/lib/utils.ts
@@ -1,6 +1,6 @@
-import { AU } from './data/bodies.ts';
-import { Time } from './epoch.ts';
-import { CelestialBody, CelestialBodyType } from './types.ts';
+import {AU} from './data/bodies.ts';
+import {Time} from './epoch.ts';
+import {CelestialBody, CelestialBodyType} from './types.ts';
 
 export function pluralize(n: number, unit: string) {
   const nAbs = Math.abs(n);
@@ -67,11 +67,3 @@ export function nameToId(name: string, shortName?: string) {
   return (shortName ?? name).replace(/\s+/g, '-').replace(/'/g, '-').toLowerCase();
 }
 
-export function currentDateSentence() {
-  const today = new Date().toLocaleDateString('en-US', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  });
-  return `Today's date is ${today}.`;
-}
